Fixes. Added passiontide.
Added passion tide, fixed sept ember days, other various fixes.

The temporal classes are a mess. The functions are overridden in the ordinariate-temporale.rb file (e.g. changes to ferial days are done in the ordinariate class), but the celebrations in the celebrations funciton need to be in the base temporale class. That needs to be fixed.

Other fixes.